modified install script to include wayland support

+++ install.sh
@@ -5,5 +5,12 @@ cp -r gnomehub@gnome-hub.github.io ~/.local/share/gnome-shell/extensions/
 gnome-extensions enable gnomehub@gnome-hub.github.io
 
 # restart gnome shell to reload the extensions
-killall -SIGQUIT gnome-shell
+if [[ -v $WAYLAND_DISPLAY ]];
+then
+    echo "wayland detected"
+    dbus-run-session -- gnome-shell --nested --wayland
+else
+    echo "x11 detected"
+    killall -SIGQUIT gnome-shell
+fi
